Create seperate test file for get index request
