+++ test/post_request.spec.js
@@ -46,23 +46,5 @@ describe('API Routes', () => {
         done();
       })
     });
-
-  });
-
-  describe('get api/v1/foods', () => {
-    it("should return all food entries in database", done =>  {
-      chai.request(server)
-      .get("/api/v1/foods")
-      .end((err, response) => {
-        response.should.have.status(200);
-        var foods = JSON.parse(response.body)
-        foods[0].should.have.property('id')
-        foods[0].should.have.property('name')
-        foods[0].should.have.property('calories')
-        foods[0].name.should.equal('goodburger')
-        foods[0].calories.should.equal(42)
-        done();
-      })
-    });
   });
 });
